Add assert to make sure testing data is correctly loaded

+++ fall-2018/capstone/homeworks/hw1/src/quicksort.h
@@ -3,7 +3,7 @@
 #include <cstddef> // size_t
 #include <iostream>
 
-using IndexType = long long;
+using IndexType = int;
 
 
 template<class T>
